requirements.txt
varint~=1.0.2
lmdb~=1.4.0
faker~=24.0.0
pycryptodome==3.20.0
